VT-491 bigdecimal migraatioskripti VT-491 Null-arvo tarkistus bigdecimal-konvertteriin VT-491 BigDecimal konvertteri morphia domainiin VT-491 Doublet muutettu BigDecimaleiksi

+++ valintaperusteet-laskenta/src/main/java/fi/vm/sade/service/valintaperusteet/service/validointi/virhe/ProsenttiosuusEpavalidiVirhe.java
@@ -1,24 +1,25 @@
 package fi.vm.sade.service.valintaperusteet.service.validointi.virhe;
 
-import fi.vm.sade.service.valintaperusteet.model.JsonViews;
+import java.math.BigDecimal;
+
 import org.codehaus.jackson.map.annotate.JsonView;
 
+import fi.vm.sade.service.valintaperusteet.model.JsonViews;
+
 /**
- * User: kwuoti
- * Date: 5.3.2013
- * Time: 9.09
+ * User: kwuoti Date: 5.3.2013 Time: 9.09
  */
 public class ProsenttiosuusEpavalidiVirhe extends Validointivirhe {
 
     @JsonView(JsonViews.Basic.class)
-    private double prosenttiosuus;
+    private BigDecimal prosenttiosuus;
 
-    public ProsenttiosuusEpavalidiVirhe(String virheviesti, double prosenttiosuus) {
+    public ProsenttiosuusEpavalidiVirhe(String virheviesti, BigDecimal prosenttiosuus) {
         super(Virhetyyppi.PROSENTTIOSUUS_EPAVALIDI, virheviesti);
         this.prosenttiosuus = prosenttiosuus;
     }
 
-    public double getProsenttiosuus() {
+    public BigDecimal getProsenttiosuus() {
         return prosenttiosuus;
     }
 }
